refactor: Comment 위치 변경 #21
